% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_edges_to_adjacency.R
\name{get_adjacency.igraph}
\alias{get_adjacency.igraph}
\title{get_adjacency.igraph}
\usage{
\method{get_adjacency}{igraph}(edges, type = c("unipartite", "bipartite"), directed = F)
}
\arguments{
\item{edges}{Can be a table which is a list pair of nodes (nodes ids are one the two first columns) a numerical third column can be associated will be the connections values.
Or an igraph object}

\item{type}{network type can be `'bipartite'` or `'unipartite'`}

\item{directed}{whether or not connections are directed (`T`) or symetrical (`F`) (default is set to `TRUE`)}
}
\value{
an adjacency/incidence matrix (data.frame) representing the network
}
\description{
An igraph object
}
\examples{

# For igraph object the usage is the same

require("igraphdata")
data("karate",package = "igraphdata")
class(karate)

get_adjacency(karate)

}
