% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextables.R
\name{get_flextable}
\alias{get_flextable}
\title{get_flextable}
\usage{
get_flextable(
  sbm,
  labels = "default",
  type = c("blockProp", "connectParam", "storedModels"),
  caption = "default",
  visualSettings = list()
)
}
\arguments{
\item{sbm}{an sbm model product of {sbm} estimation (simple or bipartite)}

\item{labels}{labels for nodes.
If it's simple sbm it should be a single character ("default" -> c("nodes")).
If sbm is bipartite a named character (names are row and col) ("default" -> c(row = 'row', col = 'col')).}

\item{type}{the type of table you want.
'blockProp' gives the block proportions.
'connectParam' gives the block connectivity.
'storedModels' gives the stored modems summary.}

\item{caption}{caption of the table. "default" gives the default title for each type.}

\item{visualSettings}{a lists of visual settings. selected_col is the color of the selected model line.
best_col is the color of the best model line and .
Default : list(selected_col = "orange", best_col = "red")}
}
\value{
Return the selected flextable
}
\description{
A fct that build a flextable from an sbm object
}
\examples{

my_sbm <- sbm::estimateBipartiteSBM(sbm::fungusTreeNetwork$fungus_tree,model = 'bernoulli')

get_flextable(my_sbm, labels = c(row = 'Fungus', col = 'Trees'),
 type = 'blockProp', caption = "default")

get_flextable(my_sbm, labels = c(row = 'Fungus', col = 'Trees'),
 type = 'connectParam', caption = "default")

get_flextable(my_sbm, labels = 'default',
type = 'storedModels', caption = "New Title")

}
