% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_is_bipartite.R
\name{is.bipartite.default}
\alias{is.bipartite.default}
\title{is.bipartite.default}
\usage{
\method{is.bipartite}{default}(object)
}
\arguments{
\item{object}{is a list that has to be checked whether or not is is bipartite}
}
\value{
TRUE if bipartite, FALSE if not
}
\description{
it's a function that says if your object (graph, sbm like or pre-sbm) is a bipartite one
}
\examples{

is.bipartite(1)

}
