% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_is_bipartite.R
\name{is.bipartite.list}
\alias{is.bipartite.list}
\title{is.bipartite.list}
\usage{
\method{is.bipartite}{list}(object)
}
\arguments{
\item{object}{is a list that has to be checked whether or not is is bipartite}
}
\value{
TRUE if bipartite, FALSE if not
}
\description{
it's a function that says if your object (graph, sbm like or pre-sbm) is a bipartite one
}
\examples{

my_net1 <- list(pop = 8, networktype = "bipartite")
is.bipartite(my_net1)

my_net2 <- list(pop = 8, networktype = "unipartite")
is.bipartite(my_net2)

}
