% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_is_bipartite.R
\name{is.bipartite}
\alias{is.bipartite}
\title{is.bipartite}
\usage{
is.bipartite(object)
}
\arguments{
\item{object}{is a list or a result of {sbm} estimation that has to be checked whether or not is is bipartite}
}
\value{
TRUE if bipartite, FALSE if not
}
\description{
it's a function that says if your object (graph, sbm like or pre-sbm) is a bipartite one
}
\examples{
my_net1 <- list(pop = 8, networktype = "bipartite")
is.bipartite(my_net1)

my_net2 <- list(pop = 8, networktype = "unipartite")
is.bipartite(my_net2)

my_sbm <- sbm::estimateBipartiteSBM(sbm::fungusTreeNetwork$fungus_tree,model = 'bernoulli')
is.bipartite(my_sbm)

}
