% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_is_bipartite.R
\name{is.bipartite.SBM}
\alias{is.bipartite.SBM}
\title{is.bipartite.SBM}
\usage{
\method{is.bipartite}{SBM}(object)
}
\arguments{
\item{object}{an sbm model from {sbm} package that has to be checked whether or not is is bipartite}
}
\value{
a Boolean, TRUE for bipartite or FALSE if else
}
\description{
it's a function that says if your sbm is bipartite or not
}
\examples{

my_sbm <- sbm::estimateBipartiteSBM(sbm::fungusTreeNetwork$fungus_tree,model = 'bernoulli')
is.bipartite(my_sbm)

}
