% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotSbm.R
\name{plotSbm.default}
\alias{plotSbm.default}
\title{plotSbm.default Method}
\usage{
\method{plotSbm}{default}(
  x,
  ordered = FALSE,
  transpose = FALSE,
  labels = NULL,
  plotOptions = list()
)
}
\arguments{
\item{x}{any object}

\item{ordered}{isn't used in default method}

\item{transpose}{isn't used in default method}

\item{labels}{isn't used in default method}

\item{plotOptions}{isn't used in default method}
}
\value{
default plot for x
}
\description{
plotSbm method for unknown object
}
