% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextables.R
\name{round_proportion}
\alias{round_proportion}
\title{round_proportion}
\usage{
round_proportion(x, digits = 2)
}
\arguments{
\item{x}{vector of proportions}

\item{digits}{number of digits, will be evaluated by round}
}
\value{
the rounded vector
}
\description{
A fct that apply round() to a proportion vector (between 0 and 1, and summing to 1). Round by digits, and adjust the last element to the summing to 1 condition
}
\examples{
options(digits = 5)
vec <- c(0.5698,0.125,0.0556,0.0365,0.1778,0.0353)
vec
sum(vec)

rounded_vec <- round_proportion(vec)
rounded_vec
sum(rounded_vec)

}
