% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visSbm.R
\name{visSbm.BipartiteSBM_fit}
\alias{visSbm.BipartiteSBM_fit}
\title{visSbm}
\usage{
\method{visSbm}{BipartiteSBM_fit}(x, labels = "default", directed = "default", settings = list())
}
\arguments{
\item{x}{: Sbm model of class `BipartiteSBM_fit`, `SimpleSBM_fit` or simple numeric `matrix`.}

\item{labels}{: labels for nodes. If it's simple sbm it should be a single character ("default" -> c("nodes")). If sbm is bipartite a named character (names are row and col) ("default" -> c(row = 'row', col = 'col')).}

\item{directed}{: Boolean indicating whether or not the network is directed by default, a asymmetrical matrix will be seen as directed.}

\item{settings}{: list of settings}
}
\value{
a visNetwork plot
}
\description{
A fct that plot a {visNetwork} plot of a adjacency matrix or an Sbm fit from the {sbm} package.
}
\details{
The list of parameters \code{plotOptions} for the matrix plot is
\itemize{
 \item{"arrow_start": }{character : "from" the arrow strat from first column and "to" the arrow start from tahe second column}
 \item{"arrows_color": }{character : color of arrows}
 \item{"row_color": }{character : color for rows in case of bipartite}
 \item{"row_shape": }{character : shape of rows according to vsiNetwork shape agrument ("triangle", "square", etc...)}
 \item{"col_color": }{character : color for columns in case of bipartite}
 \item{"col_shape": }{character : shape of columns according to vsiNetwork shape agrument ("triangle", "square", etc...)}
}
}
\examples{
data_bi <- sbm::fungusTreeNetwork$fungus_tree
my_sbm_bi <- sbm::estimateBipartiteSBM(data_bi)


visSbm(my_sbm_bi)



}
