% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotSbm.R
\name{plotSbm.default}
\alias{plotSbm.default}
\title{plotSbm.default Method}
\usage{
\method{plotSbm}{default}(
  fit,
  ordered = FALSE,
  transpose = FALSE,
  labels = NULL,
  plotOptions = list()
)
}
\arguments{
\item{fit}{: any object}

\item{ordered}{: isn't used in default method}

\item{transpose}{: isn't used in default method}

\item{labels}{: isn't used in default method}

\item{plotOptions}{: isn't used in default method

 @return plot of fit}
}
\description{
plotSbm method for unknown object
}
