% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotSbm.R
\name{plotSbm}
\alias{plotSbm}
\title{plotSbm}
\usage{
plotSbm(
  fit,
  ordered = FALSE,
  transpose = FALSE,
  labels = NULL,
  plotOptions = list()
)
}
\description{
A fct that plot a beautiful matrix from an sbm object or a network matrix it does
have suitable parameters to get the plots you want this is the generic function,
it does have one method Bipartite and one for Simple Sbm. The `fit` object need
to be construct by one of the `estimate***SBM` function from the `sbm` package.

 @param fit  : an Sbm model of class `"BipartiteSBM_fit"` or `"SimpleSBM_fit"`
 @param ordered : Boolean. Set \code{TRUE} if the matrix should be reordered (Default is \code{FALSE})
 @param transpose : Boolean. Set \code{TRUE} if you want to invert columns and rows to flatten a long matrix (Default is \code{FALSE})
 @param labels : a named list (names should be : `"col"` and `"row"`) of characters describing columns and rows component (Default is \code{NULL})
 @param plotOptions : a list providing options. See details below.
 @details The list of parameters \code{plotOptions} for the matrix plot is
 \itemize{
 \item{"showValues": }{Boolean. Set TRUE if you want to see the real values. Default value is TRUE}
 \item{"showPredictions": }{Boolean. Set TRUE if you want to see the predicted values. Default value is TRUE}
 \item{"title": }{Title in characters. Will be printed at the bottom of the matrix. Default value is NULL}
 \item{"colPred": }{Color of the predicted values, the small values will be more transparent. Default value is "red"}
 \item{"colValue": }{Color of the real values, the small values will close to white. Default value is "black"}
 }

 @return a ggplot object corresponding to the plot
}
\examples{
data_bi <- fungusTreeNetwork$fungus_tree
my_sbm_bi <- estimateBipartiteSBM(data_bi)
plotSbm(my_sbm_bi, ordered = T, transpose = T, plotOptions = list(title = 'An example Matrix'))
}
